% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_images.R
\name{download_images}
\alias{download_images}
\title{Download Images from PENTAX (TM) cameras}
\usage{
download_images(files = list_images(verbose = FALSE), output_path,
  url = "http://192.168.0.1/v1", verbose = TRUE)
}
\arguments{
\item{files}{\code{\link{list}} (with default): list with files to import}

\item{output_path}{\code{\link{character}} \bold{(required)}: character string
with a valid output path, where the downloaded files will be stored}

\item{url}{\code{\link{character}} (with default): the url for server connection}

\item{verbose}{\code{\link{logical}} (with default): enables or disables verbose mode}
}
\value{
This function returns nothing.
}
\description{
Download previously selected images. If the default is chosen all found images
are downloaded.
}
\details{
\bold{How to establish the Wi-Fi connection?}

1. Switch on the Wi-Fi of your camera \cr
2. Connect your computer with the Wi-Fi of the camera (you find the Wi-Fi password
in your camera settings: Settings >> 3 >> Wi-Fi >> Communication Info)\cr
}
\note{
This function has been tested with a Pentax K-70 (Firmware: v1.10) only! It should
theoretically work with other models providing this functionalit; however, it has been
not tested.
}
\examples{

\dontrun{
download_images(output_path = "~/Desktop/Download")

}

}
\author{
Sebastian Kreutzer
}
