% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_images.R
\name{list_images}
\alias{list_images}
\title{Show All Images on the Camera}
\usage{
list_images(url = "http://192.168.0.1/v1", verbose = TRUE)
}
\arguments{
\item{url}{\code{\link{character}} (with default): the url for the server connection}

\item{verbose}{\code{\link{logical}} (with default): enables or disables verbose mode}
}
\value{
Returns a list with the files on the camera
}
\description{
List all files found on the camera. The output of the functons is also
used for the function \code{\link{download_files}}
}
\author{
Sebastian Kreutzer
}
