% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psync.R
\name{Psync-package}
\alias{Psync-package}
\title{Downloading PENTAX (TM) Camera Images}
\description{
Some PENTAX (TM) cameras (e.g., K-1, K-70) allow a remote control and image download for via Wi-Fi.
Unfortunately, this feature appears to be so far limited to same smartphone apps. 'Psync' provides
a simple routine for selecting images and downloading them from a PENTAX (TM) camera; so far the
Wi-Fi connection has been established previously. The download is rather slow, but very convenient
if time does not matter and/or no proper cable is at hand.
}
\details{
\tabular{ll}{Package: \tab Psync\cr
 Type: \tab Package\cr
 Version: \tab 0.1.0\cr
 Date: \tab 2017-05-19 \cr
 License: \tab GPL-3 \cr}

\bold{Acknowledgement}\cr

This package could be very easily compiled thanks to the work by Claus Zielke,
who programmed the software 'Pentax-K-S2-WiFiRemote and compiled' information
on the camera WebAPI: \url{https://sourceforge.net/projects/pentaxks2wifiremote/}
}
\author{
Sebastian Kreutzer
}
\keyword{package}
